feat:建立type資料夾，將auth的type類型整合進去
